+++ server/runtime_javascript_nakama.go
@@ -9786,7 +9786,7 @@ func leaderboardToJsObject(leaderboard *api.Leaderboard) (map[string]interface{}
 }
 
 func tournamentToJsObject(tournament *api.Tournament) (map[string]interface{}, error) {
-	tournamentMap := make(map[string]interface{}, 18)
+	tournamentMap := make(map[string]interface{}, 19)
 
 	tournamentMap["id"] = tournament.Id
 	tournamentMap["title"] = tournament.Title
@@ -9821,6 +9821,7 @@ func tournamentToJsObject(tournament *api.Tournament) (map[string]interface{}, e
 		tournamentMap["endTime"] = tournament.EndTime.Seconds
 	}
 	tournamentMap["operator"] = strings.ToLower(tournament.Operator.String())
+	tournamentMap["joinRequired"] = tournament.JoinRequired
 
 	return tournamentMap, nil
 }